+++ consensus/src/context/task.rs
@@ -44,7 +44,7 @@ pub struct ContextTask {
     hardfork_state: hardforks::HardForkState,
 
     /// The current chain height.
-    chain_height: u64,
+    chain_height: usize,
     /// The top block hash.
     top_block_hash: [u8; 32],
     /// The total amount of coins generated.
